+++ packages/devtools/client/src/entries/mount/index.tsx
@@ -1,0 +1,27 @@
+import { createRoot } from 'react-dom/client';
+import { parseQuery } from 'ufo';
+import _ from 'lodash';
+import { SetupClientParams } from '@modern-js/devtools-kit';
+import { waitClientConnection } from './rpc';
+import { DevtoolsActionButton } from '@/components/Devtools/Action';
+
+// @ts-expect-error
+const { container, resourceQuery } = window._modern_js_devtools_app;
+const root = createRoot(container);
+const parsed = parseQuery(resourceQuery);
+if (
+  !_.isString(parsed.dataSource) ||
+  !_.isString(parsed.def) ||
+  !_.isString(parsed.endpoint)
+) {
+  throw new TypeError(
+    `Failed to parse client definitions of devtools: ${resourceQuery}`,
+  );
+}
+const options: SetupClientParams = {
+  dataSource: parsed.dataSource,
+  endpoint: parsed.endpoint,
+  def: JSON.parse(parsed.def),
+};
+waitClientConnection();
+root.render(<DevtoolsActionButton {...options} />);