+++ packages/generator/generator-common/src/newAction/module/index.ts
@@ -104,7 +104,7 @@ export const ModuleNewActionPluginName: Partial<
   [ActionType.Function]: {
     [ActionFunction.TailwindCSS]: 'tailwindcssPlugin',
     [ActionFunction.ModuleDoc]: 'modulePluginDoc',
-    [ActionFunction.Test]: 'testPlugin',
+    [ActionFunction.Test]: 'testingPlugin',
   },
 };