+++ packages/solutions/module-tools/src/utils/index.ts
@@ -8,5 +8,4 @@ export * from './log';
 export * from './map';
 export * from './print';
 export * from './style';
-export * from './tspath';
 export * from './outExtension';