+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @modern-js/plugin-bff
 
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/bff-core@2.42.1
+- @modern-js/create-request@2.42.1
+- @modern-js/server-utils@2.42.1
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- Updated dependencies [a3f2269]
+  - @modern-js/bff-core@2.42.0
+  - @modern-js/create-request@2.42.0
+  - @modern-js/server-utils@2.42.0
+  - @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes