+++ packages/cli/plugin-tailwind/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @modern-js/plugin-tailwindcss
 
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/runtime@2.42.1
+- @modern-js/node-bundle-require@2.42.1
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- Updated dependencies [103cf92]
+- Updated dependencies [ecc57b6]
+- Updated dependencies [d6596cb]
+- Updated dependencies [e2c899b]
+  - @modern-js/runtime@2.42.0
+  - @modern-js/node-bundle-require@2.42.0
+  - @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes