+++ packages/generator/generators/module-test-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/module-test-generator
 
+## 3.3.2
+
+### Patch Changes
+
+- @modern-js/utils@2.42.1
+
+## 3.3.1
+
+### Patch Changes
+
+- @modern-js/utils@2.42.0
+
 ## 3.3.0
 
 ### Patch Changes