+++ packages/server/bff-runtime/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/bff-runtime
 
+## 2.42.1
+
+## 2.42.0
+
 ## 2.41.0
 
 ### Patch Changes