+++ packages/server/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/server-plugin
 
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/plugin@2.42.1
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- @modern-js/plugin@2.42.0
+- @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes