+++ packages/server/plugin-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/plugin-server
 
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/server-utils@2.42.1
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- @modern-js/server-utils@2.42.0
+- @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes