+++ packages/toolkit/utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/utils
 
+## 2.42.1
+
+## 2.42.0
+
 ## 2.41.0
 
 ### Patch Changes